third_party/blink/web_tests/virtual/unload-allowed/external/wpt/html/webappapis/scripting/events/event-handler-attributes-frameset-window-expected.txt

This is a testharness.js-based test.
Found 6 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] not shadowed copy (document.body)
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed cut (document.body)
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed paste (document.body)
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed copy (document.createElement("frameset"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed cut (document.createElement("frameset"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
[FAIL] not shadowed paste (document.createElement("frameset"))
  assert_equals: window should reflect expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
